Move StructureWidget tests to separate class
