+++ tests/molara/gui/test_main_window.py
@@ -4,21 +4,18 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import numpy as np
 from molara.Gui.builder import BuilderDialog
 from molara.Gui.crystal_dialog import CrystalDialog
 from molara.Gui.main_window import MainWindow
 from molara.Gui.measuring_tool_dialog import MeasurementDialog
-from molara.Gui.structure_widget import BUILDER, MEASUREMENT, StructureWidget
+from molara.Gui.structure_widget import StructureWidget
 from molara.Gui.trajectory_dialog import TrajectoryDialog
 from molara.Gui.ui_form import Ui_MainWindow
 from molara.Structure.crystal import Crystal
 from molara.Structure.crystals import Crystals
 from molara.Structure.molecule import Molecule
 from molara.Structure.molecules import Molecules
-from numpy.testing import assert_array_almost_equal
-from PySide6.QtCore import QEvent, QPoint, Qt
-from PySide6.QtGui import QAction, QMouseEvent, QSurfaceFormat
+from PySide6.QtGui import QAction, QSurfaceFormat
 from PySide6.QtWidgets import QApplication, QMenu, QMenuBar
 
 if TYPE_CHECKING:
@@ -53,8 +50,6 @@ class WorkaroundTestMainWindow:
         """Run all tests."""
         self.test_init()
         self.test_ui()
-        self.test_structure_widget()
-        self.test_set_view_to_axes()
         self.test_export_image_dialog()
         self.test_show_builder_dialog()
         self.test_show_crystal_dialog()
@@ -160,104 +155,6 @@ class WorkaroundTestMainWindow:
         """Clean up the test."""
         self.window.close()
         QApplication.instance().shutdown() if QApplication.instance() is not None else None
-
-    def test_structure_widget(self) -> None:
-        """Write test code to verify the behavior of the structure_widget property."""
-        structure_widget = self.window.structure_widget
-        structure_widget.toggle_unit_cell_boundaries()
-        testargs = ["molara", "examples/xyz/pentane.xyz"]
-        with mock.patch.object(sys, "argv", testargs):
-            self.window.show_init_xyz()
-        assert structure_widget.draw_bonds
-        assert structure_widget.bonds
-        self.window.structure_customizer_dialog.toggle_bonds()
-        assert not structure_widget.bonds
-        assert not structure_widget.draw_bonds
-        assert structure_widget is not None
-
-        # Test mouse moves and clicks
-        self.qtbot.mousePress(structure_widget, Qt.LeftButton, pos=QPoint(50, 50))
-
-        # Simulate mouse move events to rotate the structure
-        self.qtbot.mouseMove(structure_widget, QPoint(60, 60))
-        self.qtbot.mouseMove(structure_widget, QPoint(70, 70))
-
-        # Simulate a left mouse button release event to end rotation
-        self.qtbot.mouseRelease(structure_widget, Qt.LeftButton, pos=QPoint(70, 70))
-
-        self.qtbot.mousePress(structure_widget, Qt.RightButton, pos=QPoint(50, 50))
-
-        # Simulate mouse move events to rotate the structure
-        self.qtbot.mouseMove(structure_widget, QPoint(60, 60))
-        self.qtbot.mouseMove(structure_widget, QPoint(70, 70))
-
-        # Simulate a left mouse button release event to end rotation
-        self.qtbot.mouseRelease(structure_widget, Qt.RightButton, pos=QPoint(70, 70))
-
-        # Test toggle axes:
-        structure_widget.toggle_axes()
-        assert structure_widget.draw_axes
-        structure_widget.toggle_axes()
-        assert not structure_widget.draw_axes
-
-        # Test measurement select sphere
-        event = QMouseEvent(
-            QEvent.MouseButtonPress,  # Event type
-            QPoint(50, 50),  # Position
-            Qt.LeftButton,  # Button
-            Qt.LeftButton,  # Buttons (pressed buttons)
-            Qt.NoModifier,  # Modifiers (keyboard modifiers)
-        )
-        structure_widget.update_selected_atoms(MEASUREMENT, event)
-
-        measurement_selected_spheres = structure_widget.measurement_selected_spheres
-        structure_widget.unselect_all_atoms()
-        for sphere_id in measurement_selected_spheres:
-            assert sphere_id == -1
-        id1, id2, id3, id4 = 2, 1, 0, 3
-        structure_widget.exec_select_sphere(id1, measurement_selected_spheres)
-        assert measurement_selected_spheres[0] == id1
-        assert_array_almost_equal(
-            structure_widget.structures[0].drawer.atom_colors[id1],
-            structure_widget.new_sphere_colors[0],
-        )
-        structure_widget.exec_select_sphere(id2, measurement_selected_spheres)
-        assert measurement_selected_spheres[1] == id2
-        assert_array_almost_equal(
-            structure_widget.structures[0].drawer.atom_colors[id2],
-            structure_widget.new_sphere_colors[1],
-        )
-        structure_widget.exec_select_sphere(id3, measurement_selected_spheres)
-        assert measurement_selected_spheres[2] == id3
-        assert_array_almost_equal(
-            structure_widget.structures[0].drawer.atom_colors[id3],
-            structure_widget.new_sphere_colors[2],
-        )
-        structure_widget.exec_select_sphere(id4, measurement_selected_spheres)
-        assert measurement_selected_spheres[3] == id4
-        assert_array_almost_equal(
-            structure_widget.structures[0].drawer.atom_colors[id4],
-            structure_widget.new_sphere_colors[3],
-        )
-
-        # Test builder select sphere
-        structure_widget.update_selected_atoms(BUILDER, event)
-
-    def test_set_view_to_axes(self) -> None:
-        """Test setting the rotation of the camera. See also test_set_rotation in test_camera.py."""
-        structure_widget = self.window.structure_widget
-        camera = structure_widget.camera
-        quaternion_x = [0.0, 0.0, 0.0, 1.0]
-        quaternion_y = [0.0, 0.0, np.sqrt(2) / 2, np.sqrt(2) / 2]
-        quaternion_z = [0.0, np.sqrt(2) / 2, 0.0, np.sqrt(2) / 2]
-        # quaternion_y = [0.0, 0.0, 0.7071067690849304, 0.7071067690849304]
-        # quaternion_z = [0.0, 0.7071067690849304, 0.0, 0.7071067690849304]
-        structure_widget.set_view_to_axis("x")
-        assert_array_almost_equal(camera.rotation.tolist(), quaternion_x)
-        structure_widget.set_view_to_axis("y")
-        assert_array_almost_equal(camera.rotation.tolist(), quaternion_y)
-        structure_widget.set_view_to_axis("z")
-        assert_array_almost_equal(camera.rotation.tolist(), quaternion_z)
 
     def test_export_image_dialog(self) -> None:
         """Write test code to verify the behavior of export_image_dialog property."""
